app/schemas/project.py
from pydantic import BaseModel
from typing import Optional
class ProjectBase(BaseModel):
    name: str
    description: Optional[str] = None
class ProjectCreate(ProjectBase):
    pass
class ProjectOut(ProjectBase):
    id: int
    owner_id: int | None = None
    class Config:
        from_attributes = True